fix axis length data type

+++ src/main/java/jp/ac/aiit/pbl/format/sep2020/SemiMinorAxisLength.java
@@ -2,13 +2,13 @@ package jp.ac.aiit.pbl.format.sep2020;
 
 public class SemiMinorAxisLength {
 
-    private Double value;
+    private long value;
 
     public SemiMinorAxisLength(String axis){
         this.value = AxisLength.getById(axis).getLength();
     }
 
-    public Double value(){
+    public long value(){
         return value;
     }
 }
